requirements.txt
absl-py==1.0.0
accelerate==0.5.1
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.10
colorama==0.4.4
google-auth==2.5.0
google-auth-oauthlib==0.4.6
grpcio==1.43.0
idna==3.3
importlib-metadata==4.10.1
install==1.3.5
kornia==0.6.2
Markdown==3.3.6
numpy>=1.20.1
oauthlib==3.1.1
packaging==21.3
Pillow==9.0.0
protobuf==3.19.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.7
python-math==0.0.1
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.0
rsa==4.8
six==1.16.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
timm==0.5.4
torch==1.10.1
torchvision==0.11.2
tqdm==4.62.3
typing-extensions==4.0.1
urllib3==1.26.8
Werkzeug==2.0.2
zipp==3.7.0
